Ignore pen inputs on canvas when activating the window
It is annoying to click or tap on the window to focus it and end
up accidentally drawing a dot onto the canvas. This change makes
Drawpile work more like Photoshop, etc. that discard inputs on
window activation.

The approach of a timer and some extra state management is used
because better alternatives do not seem to exist given these
limitations:

1. Qt has no API to tell it to not send mouse events that trigger
   a window activation on macOS (QTBUG-83713), and anyway, tablet
   events that cause window activation are also sent on Windows;
2. Qt has no API to inspect/purge the internal event queue, and
   does not seem to alway receive or process the whole OS event
   queue at the time the window activation event is received, so
   it appears to be impossible to just push a low-priority signal
   event to the queue;
3. Qt does not mark input events that caused a window activation
   so that they can be discarded;
4. Something causes events to be sent out-of-order (old release
   events before new press events), duplicated (tablet triggering
   both mouse and tablet events, sometimes with a delay between
   them), or lost (release events), so state management is needed
   to handle these edge cases.

This approach could be extended to the entire UI, but it is normal
for e.g. a click on a text widget to also focus the widget along
with the window (at least on macOS).