Use sqlite instead of QSortFilterProxyModel for sorting
This reduces startup time by ~650ms with the default set of
built-in brushes whilst maintaining the desired sorting behaviour
in the brush palette dock.

It is not clear what is the problem with QSortFilterProxyModel;
sorting should not take 3.4ms per item. There are some reports here
and there about this problem, but without any obvious cause or
solution.

QSortFilterProxyModel is still used for filtering since it is
easier than tearing it all out completely.