Report all paint engine crashes and make them plausibly recoverable
Since the cached view state is not updated until all queued
messages are processed, if there is a panic when processing any
message, it is possible to attempt a recovery with a reasonable
chance of success by restarting with the cached view state.

In the future, clients connected to a server should re-request a
fresh view state from the server, but I was not able to make sense
of the login flow to see how to do that right now in a reasonable
period of time.

Previously, panics that happened in the paint engine from outside
of the two message queue entrypoints were never reported. Now,
crashes will always be reported whenever something tries to
interact with the paint engine after it crashed. Ideally, crash
notification would be sent immediately, but I was not confident
that the callback would be thread-safe in that case, so opted to
take the more conservative approach at the moment of recovering the
error from the sender thread.