Default to HTTPS when adding list servers via join dialog
Also allows adding list servers that use HTTP authentication or
run on a non-standard port here, to match what you can do in the
preferences.

This design really needs a rethink. This input box is what one uses
to join a Drawpile server directly, but then this code also lets
you add a list server by putting in a list server address, but a
Drawpile server and a list server are two different things. So
anyone who puts in a Drawpile server here to join a session, then
tries to “add” it so they can see the sessions on the server, will
only experience confusion as nothing gets added unless someone is
running a list server and a Drawpile server on the same domain.